+++ .gitignore
@@ -7,8 +7,7 @@ glide.lock
 # binary files
 *.log
 *.exe
-edgexsecurity
 
-core/edgexproxy
-core/edgexproxy.exe
+cmd/edgexproxy/*.exe
+cmd/edgexproxy/edgexproxy
 *.sum
