--- core/accessToken.json
@@ -1,4 +1,0 @@
-{
- "User": "aaamd",
- "Token": "yIPD6phz1zhFiF7PwRa7gDV4GP0YVSAe"
-}
