--- core/res/resp-init.json
@@ -1,9 +1,0 @@
-{
-    "keys": [
-      "69be8924eea8b553755d090f6f38a7ebab24d17b8a6a515ae5f11cb13872b532"
-    ],
-    "keys_base64": [
-      "ab6JJO6otVN1XQkPbzin66sk0XuKalFa5fEcsThytTI="
-    ],
-    "root_token": "232d6e0a-3a09-1090-aa08-b303f28c05e1"
-}
